+++ testdmd.py
@@ -8,7 +8,14 @@ import util
 import subprocess as SP
 import os
 import copy
+import parallel as parallel_mod
 
+parallel = parallel_mod.parallelInstance
+
+if parallel.isRankZero():
+    print 'To test fully, remember to do both:'
+    print '    1) python testdmd.py'
+    print '    2) mpiexec -n <# procs> python testdmd.py\n'
 
 class TestDMD(unittest.TestCase):
     """ Test all the DMD class methods 
@@ -34,14 +41,22 @@ class TestDMD(unittest.TestCase):
         self.snapPathList = []
        
         # Generate modes if we are on the first processor
-        # A random matrix of data (#cols = #snapshots)
-        self.snapMat = N.mat(N.random.random((self.numStates, self.\
-            numSnaps)))
-        
-        for snapNum in range(self.numSnaps):
-            util.save_mat_text(self.snapMat[:,snapNum], self.snapPath %\
-                snapNum)
-            self.snapPathList.append(self.snapPath % snapNum) 
+        if parallel.isRankZero():
+            # A random matrix of data (#cols = #snapshots)
+            self.snapMat = N.mat(N.random.random((self.numStates, self.\
+                numSnaps)))
+            
+            for snapNum in range(self.numSnaps):
+                util.save_mat_text(self.snapMat[:,snapNum], self.snapPath %\
+                    snapNum)
+                self.snapPathList.append(self.snapPath % snapNum) 
+        else:
+            self.snapPathList=None
+            self.snapMat = None
+        if parallel.isDistributed():
+            self.snapPathList = parallel.comm.bcast(self.snapPathList, 
+                root=0)
+            self.snapMat = parallel.comm.bcast(self.snapMat, root=0)
 
         # Do direct DMD decomposition on all processors
         U, Sigma, W = util.svd(self.snapMat[:,:-1])
@@ -62,62 +77,69 @@ class TestDMD(unittest.TestCase):
                 i])).real
 
         # Generate modes if we are on the first processor
-        for i in xrange(self.ritzVecsTrue.shape[1]):
-            util.save_mat_text(self.ritzVecsTrue[:,i], self.trueModePath %\
-                (i+1))
+        if parallel.isRankZero():
+            for i in xrange(self.ritzVecsTrue.shape[1]):
+                util.save_mat_text(self.ritzVecsTrue[:,i], self.trueModePath %\
+                    (i+1))
 
     def tearDown(self):
-        SP.call(['rm -rf files_modaldecomp_test/*'],shell=True)
+        parallel.sync()
+        if parallel.isRankZero():
+            SP.call(['rm -rf files_modaldecomp_test/*'],shell=True)
+        parallel.sync()
 
     def test_init(self):
         """Test arguments passed to the constructor are assigned properly"""
         # Get default data member values
         # Set verbose to false, to avoid printing warnings during tests
         
-        def getDataMembersDefault():
-            return {'save_mat': util.save_mat_text, 'load_mat': util.\
-            load_mat_text, 'pod': None, 'verbose':\
-            False, 'fieldOperations': FieldOperations(load_field=None, 
-            save_field=None, inner_product=None, maxFields=2,
+        dataMembersDefault = {'save_mat': util.save_mat_text, 'load_mat': util.\
+            load_mat_text, 'pod': None, 'parallel': parallel_mod.parallelInstance,\
+            'verbose': False, 'fieldOperations': FieldOperations(load_field=None, 
+            save_field=None, inner_product=None, maxFieldsPerNode=2,
             verbose=False)}
         
         self.assertEqual(util.get_data_members(DMD(verbose=False)), \
-            getDataMembersDefault())
+            dataMembersDefault)
 
         def my_load(fname): pass
         myDMD = DMD(load_field=my_load, verbose=False)
-        dataMembersModified = getDataMembersDefault()
+        dataMembersModified = copy.deepcopy(dataMembersDefault)
         dataMembersModified['fieldOperations'].load_field = my_load
         self.assertEqual(util.get_data_members(myDMD), dataMembersModified)
         
         myDMD = DMD(load_mat=my_load, verbose=False)
-        dataMembersModified = getDataMembersDefault()
+        dataMembersModified = copy.deepcopy(dataMembersDefault)
         dataMembersModified['load_mat'] = my_load
         self.assertEqual(util.get_data_members(myDMD), dataMembersModified)
 
         def my_save(data,fname): pass 
         myDMD = DMD(save_field=my_save, verbose=False)
-        dataMembersModified = getDataMembersDefault()
+        dataMembersModified = copy.deepcopy(dataMembersDefault)
         dataMembersModified['fieldOperations'].save_field = my_save
         self.assertEqual(util.get_data_members(myDMD), dataMembersModified)
         
         myDMD = DMD(save_mat=my_save, verbose=False)
-        dataMembersModified = getDataMembersDefault()
+        dataMembersModified = copy.deepcopy(dataMembersDefault)
         dataMembersModified['save_mat'] = my_save
         self.assertEqual(util.get_data_members(myDMD), dataMembersModified)
         
         def my_ip(f1, f2): pass
         myDMD = DMD(inner_product=my_ip, verbose=False)
-        dataMembersModified = getDataMembersDefault()
+        dataMembersModified = copy.deepcopy(dataMembersDefault)
         dataMembersModified['fieldOperations'].inner_product = my_ip
         self.assertEqual(util.get_data_members(myDMD), dataMembersModified)
 
-        maxFields = 500
-        myDMD = DMD(maxFields=maxFields, verbose=False)
-        dataMembersModified = getDataMembersDefault()
-        dataMembersModified['fieldOperations'].maxFields = maxFields
+        maxFieldsPerNode = 500
+        myDMD = DMD(maxFieldsPerNode=maxFieldsPerNode, verbose=False)
+        dataMembersModified = copy.deepcopy(dataMembersDefault)
+        dataMembersModified['fieldOperations'].maxFieldsPerNode =\
+            maxFieldsPerNode
+        dataMembersModified['fieldOperations'].maxFieldsPerProc =\
+            maxFieldsPerNode * parallel.getNumNodes() / parallel.\
+            getNumProcs()
         self.assertEqual(util.get_data_members(myDMD), dataMembersModified)
-
+       
 
     def test_compute_decomp(self):
         """ 
@@ -132,35 +154,42 @@ class TestDMD(unittest.TestCase):
         ritzValsPath = 'files_modaldecomp_test/dmd_ritzvals.txt'
         modeNormsPath = 'files_modaldecomp_test/dmd_modeenergies.txt'
         buildCoeffPath = 'files_modaldecomp_test/dmd_buildcoeff.txt'
-        
-        for sharedMemLoad in [True, False]:
-            for sharedMemInnerProduct in [True, False]:
-                self.dmd.compute_decomp(self.snapPathList, sharedMemLoad=\
-                    sharedMemLoad, sharedMemInnerProduct=sharedMemInnerProduct)
-                self.dmd.save_decomp(ritzValsPath, modeNormsPath, buildCoeffPath)
-               
-                # Test that matrices were correctly computed
-                N.testing.assert_array_almost_equal(self.dmd.ritzVals, self.\
-                    ritzValsTrue, decimal=tol)
-                N.testing.assert_array_almost_equal(self.dmd.buildCoeff, self.\
-                    buildCoeffTrue, decimal=tol)
-                N.testing.assert_array_almost_equal(self.dmd.modeNorms, self.\
-                    modeNormsTrue, decimal=tol)
 
-                # Test that matrices were correctly stored
-                ritzValsLoaded = N.array(util.load_mat_text(ritzValsPath,
-                    isComplex=True)).squeeze()
-                buildCoeffLoaded = util.load_mat_text(buildCoeffPath,isComplex=\
-                    True)
-                modeNormsLoaded = N.array(util.load_mat_text(modeNormsPath).\
-                    squeeze())
+        self.dmd.compute_decomp(self.snapPathList)
+        self.dmd.save_decomp(ritzValsPath, modeNormsPath, buildCoeffPath)
+       
+        # Test that matrices were correctly computed
+        N.testing.assert_array_almost_equal(self.dmd.ritzVals, self.\
+            ritzValsTrue, decimal=tol)
+        N.testing.assert_array_almost_equal(self.dmd.buildCoeff, self.\
+            buildCoeffTrue, decimal=tol)
+        N.testing.assert_array_almost_equal(self.dmd.modeNorms, self.\
+            modeNormsTrue, decimal=tol)
 
-                N.testing.assert_array_almost_equal(ritzValsLoaded, self.\
-                    ritzValsTrue, decimal=tol)
-                N.testing.assert_array_almost_equal(buildCoeffLoaded, self.\
-                    buildCoeffTrue, decimal=tol)
-                N.testing.assert_array_almost_equal(modeNormsLoaded, self.\
-                    modeNormsTrue, decimal=tol)
+        # Test that matrices were correctly stored
+        if parallel.isRankZero():
+            ritzValsLoaded = N.array(util.load_mat_text(ritzValsPath,isComplex=\
+                True)).squeeze()
+            buildCoeffLoaded = util.load_mat_text(buildCoeffPath,isComplex=True)
+            modeNormsLoaded = N.array(util.load_mat_text(modeNormsPath).\
+                squeeze())
+        else:   
+            ritzValsLoaded = None
+            buildCoeffLoaded = None
+            modeNormsLoaded = None
+
+        if parallel.isDistributed():
+            ritzValsLoaded = parallel.comm.bcast(ritzValsLoaded,root=0)
+            buildCoeffLoaded = parallel.comm.bcast(buildCoeffLoaded,root=0)
+            modeNormsLoaded = parallel.comm.bcast(modeNormsLoaded,
+                root=0)
+
+        N.testing.assert_array_almost_equal(ritzValsLoaded, self.\
+            ritzValsTrue, decimal=tol)
+        N.testing.assert_array_almost_equal(buildCoeffLoaded, self.\
+            buildCoeffTrue, decimal=tol)
+        N.testing.assert_array_almost_equal(modeNormsLoaded, self.\
+            modeNormsTrue, decimal=tol)
 
     def test_compute_modes(self):
         """
@@ -172,30 +201,29 @@ class TestDMD(unittest.TestCase):
         modePath ='files_modaldecomp_test/dmd_mode_%03d.txt'
         self.dmd.buildCoeff = self.buildCoeffTrue
         modeNumList = list(N.array(range(self.numSnaps-1))+self.indexFrom)
-        
-        for sharedMemLoad in [True, False]:
-            for sharedMemSave in [True, False]:
-        
-                self.dmd.compute_modes(modeNumList, modePath, indexFrom=self.\
-                    indexFrom, snapPaths=self.snapPathList, sharedMemLoad=\
-                    sharedMemLoad, sharedMemSave=sharedMemSave)
-               
-                # Load all snapshots into matrix
-                modeMat = N.mat(N.zeros((self.numStates, self.numSnaps-1)), 
-                    dtype=complex)
-                for i in range(self.numSnaps-1):
-                    modeMat[:,i] = util.load_mat_text(modePath % (i+self.\
-                        indexFrom), isComplex=True)
-                N.testing.assert_array_almost_equal(modeMat,self.ritzVecsTrue, 
-                    decimal=tol)
+        self.dmd.compute_modes(modeNumList, modePath, indexFrom=self.indexFrom, 
+            snapPaths=self.snapPathList)
+       
+        # Load all snapshots into matrix
+        if parallel.isRankZero():
+            modeMat = N.mat(N.zeros((self.numStates, self.numSnaps-1)), dtype=\
+                complex)
+            for i in range(self.numSnaps-1):
+                modeMat[:,i] = util.load_mat_text(modePath % (i+self.indexFrom),
+                    isComplex=True)
+        else:
+            modeMat = None
+        if parallel.isDistributed():
+            modeMat = parallel.comm.bcast(modeMat,root=0)
+        N.testing.assert_array_almost_equal(modeMat,self.ritzVecsTrue, decimal=\
+            tol)
 
-                vandermondeMat = N.fliplr(N.vander(self.ritzValsTrue,self.\
-                    numSnaps-1))
-                N.testing.assert_array_almost_equal(self.snapMat[:,:-1], self.\
-                    ritzVecsTrue * vandermondeMat, decimal=tol)
+        vandermondeMat = N.fliplr(N.vander(self.ritzValsTrue,self.numSnaps-1))
+        N.testing.assert_array_almost_equal(self.snapMat[:,:-1], self.\
+            ritzVecsTrue * vandermondeMat, decimal=tol)
 
-                util.save_mat_text(vandermondeMat, 'files_modaldecomp_test/' +\
-                    'dmd_vandermonde.txt')
+        util.save_mat_text(vandermondeMat, 'files_modaldecomp_test/' +\
+            'dmd_vandermonde.txt')
 
 if __name__=='__main__':
     unittest.main(verbosity=2)
